exo 2 selection sort + change int into luint

+++ 2-selection_sort.c
@@ -2,7 +2,7 @@
 #include "sort.h"
 void selection_sort(int *array, size_t size)
 {
-	int i, min, j, tmp;
+	long unsigned int i, min, j, tmp;
 
 	for (i = 0; i < (size - 1); i++)
 	{
